REFACTOR: MVC 패턴에 맞춰 패키지 구조 수정
