+++ src/main/java/racingcar/controller/Generator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.controller;
 
 import java.util.ArrayList;
 import java.util.List;
