+++ src/test/java/racingcar/controller/GeneratorTest.java
@@ -1,10 +1,11 @@
-package racingcar;
+package racingcar.controller;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.AssertionsForClassTypes.catchThrowable;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import racingcar.controller.Generator;
 import racingcar.model.Car;
 
 class GeneratorTest {
