+++ src/test/java/racingcar/model/RaceTest.java
@@ -1,10 +1,11 @@
-package racingcar.controller;
+package racingcar.model;
 
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import racingcar.model.Car;
+import racingcar.model.Race;
 
 class RaceTest {
